Fix build error from removed method 'CreationContext()'
The V8 method `CreationContext()` was deprecated in the version that
comes with Node.js v16, and was definitely removed starting from Node.js
v19.

Instead, the newer `GetCreationContext()` should be used; it returns a
Maybe type which can be unwrapped.

This patch uses `FromMaybe(default_value)` to unwrap the Maybe with a
default-constructed context, which is exactly the behavior that the old
`CreationContext()` had.